solo xp fix, enlightenment skill credit fixes, enlighten and aug cost adjustments, pvp config adjustments

+++ Source/ACE.Database/Models/World/WeeniePropertiesCreateList.cs
@@ -36,7 +36,7 @@ public partial class WeeniePropertiesCreateList
     /// <summary>
     /// Palette Color of Object
     /// </summary>
-    public sbyte Palette { get; set; }
+    public int Palette { get; set; }
 
     /// <summary>
     /// Shade of Object&apos;s Palette
